created the GameRunnerService.kt which deals with starting the games in an executorService, some more general code clean up

+++ src/main/kotlin/com/kj/dom/model/response/AdMessageResponse.kt
@@ -13,7 +13,7 @@ data class AdMessageResponse(
     fun toModel(): AdMessage = AdMessage(
         adId = adId,
         message = message,
-        reward = reward,
+        reward = reward.toInt(),
         expiresIn = expiresIn,
         encrypted = encrypted,
         probability = probability,
